% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateCPPGlobal}
\alias{calculateCPPGlobal}
\title{Calculated the global changepoint probabilities.}
\usage{
calculateCPPGlobal(prob.cps)
}
\arguments{
\item{prob.cps}{Node-specific changepoint probabilities, a NumNodes by
NumTimepoints matrix.}
}
\value{
A matrix of length 1 by NumTimepoints, containing the global
changepoint probabilities.
}
\description{
This function calculates the global probability of a changepoint at each
measured timepoint, using the node-specific probabilities.
}
\author{
Frank Dondelinger
}
\seealso{
\code{\link{calculateCPProbabilities}}
}

