% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateResidual.R
\name{calculateResidual}
\alias{calculateResidual}
\title{Calculates the squared residual between the target data and predicted data}
\usage{
calculateResidual(x, y, delta2)
}
\arguments{
\item{x}{Observation data.}

\item{y}{Target data.}

\item{delta2}{Delta squared parameter (signal-to-noise).}
}
\value{
Returns the squared prediction residual.
}
\description{
Wraps the cpp_calculate_residual which is implemented in C++.
}
\author{
Di Lu, Gordon McDonald
}
