% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePx.R
\name{computePx}
\alias{computePx}
\title{Compute projection matrix.}
\usage{
computePx(len, x, delta2)
}
\arguments{
\item{len}{Delimiting breakpoints.}

\item{x}{The observations of x in the corresponding state.}

\item{delta2}{Signal-to-noise ratio hyperparameter.}
}
\value{
The projection matrix.
}
\description{
This function computes the projection matrix that is needed for calculation
of the likelihood.
}
\author{
Sophie Lebre
}
\references{
For more information about the hyperparameters and the
functional form of the likelihood, see:

Dondelinger et al. (2012), "Non-homogeneous dynamic Bayesian networks with
Bayesian regularization for inferring gene regulatory networks with
gradually time-varying structure", Machine Learning.
}
\seealso{
\code{\link{CalculateLikelihoodRatio}}
}

