% Generated by roxygen2 (4.0.2): do not edit by hand
\name{NetworkProbExp}
\alias{NetworkProbExp}
\title{Calculates the prior probability of the network using the exponential prior.}
\usage{
NetworkProbExp(network.info)
}
\arguments{
\item{network.info}{Network information collected using the function
\code{\link{CollectNetworkInfo}}}
}
\value{
Returns the log prior probability of the network segments.
}
\description{
This function calculates the log prior probability of the network structure.
It uses the exponential information sharing prior.
}
\author{
Frank Dondelinger
}
\references{
For information about the exponential information sharing prior,
see:

Husmeier et al. (2010), "Inter-time segment information sharing for
non-homogeneous dynamic Bayesian networks", NIPS.

Dondelinger et al. (2012), "Non-homogeneous dynamic Bayesian networks with
Bayesian regularization for inferring gene regulatory networks with
gradually time-varying structure", Machine Learning.
}
\seealso{
\code{\link{NetworkRatioExp}}, \code{\link{CalculatePriorRatio}}
}

