% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposalTuning.R
\name{proposalTuning}
\alias{proposalTuning}
\title{Tune the proposal width for betas.}
\usage{
proposalTuning(acceptRate, hyper.proposals)
}
\arguments{
\item{acceptRate}{Current acceptance rate.}

\item{hyper.proposals}{Current proposal width.}
}
\value{
Returns the new proposal width.
}
\description{
This function adjusts the proposal width for the beta hyperparameter(s) of
the exponential information sharing prior, so that the acceptance rate is
close to 0.25.
}
\author{
Frank Dondelinger
}

