% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDelta2.R
\name{sampleDelta2}
\alias{sampleDelta2}
\title{Sample delta squared.}
\usage{
sampleDelta2(pos, x, q, B, S, sig2, alphad2, betad2)
}
\arguments{
\item{pos}{The current segment.}

\item{x}{Data,}

\item{q}{Number of nodes.}

\item{B}{Regression coefficients.}

\item{S}{Network structure.}

\item{sig2}{Sigma squared.}

\item{alphad2}{Gamma prior hyperparameter.}

\item{betad2}{Gamma prior hyperparameter.}
}
\value{
New sample of delta squared.
}
\description{
This function samples the signal-to-noise hyperparameter delta squared.
}
\author{
Sophie Lebre
}
\references{
For details of the sampling, see:

Dondelinger et al. (2012), "Non-homogeneous dynamic Bayesian networks with
Bayesian regularization for inferring gene regulatory networks with
gradually time-varying structure", Machine Learning.
}

