% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSig2.R
\name{sampleSig2}
\alias{sampleSig2}
\title{Sample initial sigma squared.}
\usage{
sampleSig2(y, x, v0, gamma0, delta2)
}
\arguments{
\item{y}{Target data.}

\item{x}{Prediction data.}

\item{v0}{Inverse gamma prior hyperparameter.}

\item{gamma0}{Inverse gamma prior hyperparameter.}

\item{delta2}{Hyperparameter.}
}
\value{
The sampled sigma squared values.
}
\description{
This function samples the initial values for the sigma squared variance from
the inverse gamma prior.
}
\references{
For more information about the model, see:

Dondelinger et al. (2012), "Non-homogeneous dynamic Bayesian networks with
Bayesian regularization for inferring gene regulatory networks with
gradually time-varying structure", Machine Learning.
}
\author{
Sophie Lebre
}
