% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperParms.R
\name{HyperParms}
\alias{HyperParms}
\title{Sets up initial values of hyperparameters.}
\usage{
HyperParms(options)
}
\arguments{
\item{options}{MCMC settings, possibly from \code{\link{defaultOptions}}.}
}
\value{
Settings for the HYPERvar variable: \item{cD}{Proportion of
changepoint moves proposed.} \item{alphaD}{Prior settings for the number of
changepoints.} \item{betaD}{Prior settings for the number of changepoints.}
\item{c}{Ratio of changepoint birth/death moves proposed.} \item{v0}{Prior
settings for the sigma squared parameters.} \item{gamma0}{Prior settings for
the sigma squared parameters.} \item{alphad2}{Prior settings for the
signal-to-noise ratio delta squared.} \item{betad2}{Prior settings for the
signal-to-noise ratio delta squared.} \item{alphalbd}{Prior settings for the
number of transcription factors.} \item{betalbd}{Prior settings for the
number of transcription factors.}
}
\description{
This function initialises the variable HYPERvar with values for the various
hyperparameters in the model.
}
\author{
Sophie Lebre

Frank Dondelinger
}

