% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psrf_check}
\alias{psrf_check}
\title{Check the potential scale reduction factors for all parameters (edges).}
\usage{
psrf_check(params, q, k_max, num_it)
}
\arguments{
\item{params}{Matrix of parameters.}

\item{q}{Number of nodes.}

\item{k_max}{Number of segments.}

\item{num_it}{Number of iterations/samples.}
}
\value{
Returns the highest PSRF value.
}
\description{
This function treats the edges of the network as parameters, calculates
their potential scale reduction factors and returns the highest value.
}
\author{
Frank Dondelinger
}
\references{
Gelman and Rubin (1992) Inference from iterative simulation
using multiple sequences, Statistical Science.
}
\seealso{
\code{\link{psrf}}, \code{\link{psrf_check_hyper}}
}

