% Generated by roxygen2 (4.0.2): do not edit by hand
\name{psrf_check_hyper}
\alias{psrf_check_hyper}
\title{Checks the potential scale reduction factor for the hyperparameters.}
\usage{
psrf_check_hyper(params, num_it)
}
\arguments{
\item{params}{Matrix of hyperparameters.}

\item{num_it}{Number of iterations/samples.}
}
\value{
Returns the maximum PSRF value.
}
\description{
This function checks the potential scale reduction factors for the
hyperparameters of the information sharing priors.
}
\author{
Frank Dondelinger
}
\references{
Gelman and Rubin (1992) Inference from iterative simulation
using multiple sequences, Statistical Science.
}
\seealso{
\code{\link{psrf}}, \code{\link{psrf_check}}
}

