% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataTS.R
\name{readDataTS}
\alias{readDataTS}
\title{Read target data.}
\usage{
readDataTS(data, posI, t0, tf, m, n)
}
\arguments{
\item{data}{Input data matrix to read.}

\item{posI}{Position of interest.}

\item{t0}{First timepoint.}

\item{tf}{Last timepoint.}

\item{m}{Number of repetitions.}

\item{n}{Number of timepoints.}
}
\value{
Returns the target data.
}
\description{
This function reads in the target data.
}
\author{
Sophie Lebre
}
\seealso{
\code{\link{buildXY}}
}

