% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBinit.R
\name{sampleBinit}
\alias{sampleBinit}
\title{Sample initial regression coefficients.}
\usage{
sampleBinit(Si, sig2, delta2, X, q)
}
\arguments{
\item{Si}{Network structure.}

\item{sig2}{Sigma squared.}

\item{delta2}{Signal-to-noise ratio hyperparameter.}

\item{X}{Input data.}

\item{q}{Number of nodes.}
}
\value{
Returns a vector of regression coefficients.
}
\description{
This function samples the initial regression coefficients for the networks.
}
\author{
Sophie Lebre
}
\references{
For details of the regression model, see:

Dondelinger et al. (2012), "Non-homogeneous dynamic Bayesian networks with
Bayesian regularization for inferring gene regulatory networks with
gradually time-varying structure", Machine Learning.
}

