% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBxy.R
\name{sampleBxy}
\alias{sampleBxy}
\title{Sample regression coefficients.}
\usage{
sampleBxy(xi, y, Sig2, delta2)
}
\arguments{
\item{xi}{Response data.}

\item{y}{Target data.}

\item{Sig2}{Sigma squared.}

\item{delta2}{Signal-to-noise hyperparameter.}
}
\value{
The regression parameters.
}
\description{
This function samples the regression coefficients given the current state of
the MCMC simulation.
}
\author{
Sophie Lebre
}
\references{
For details of the regression model, see:

Dondelinger et al. (2012), "Non-homogeneous dynamic Bayesian networks with
Bayesian regularization for inferring gene regulatory networks with
gradually time-varying structure", Machine Learning.
}

