% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sampleK}
\alias{sampleK}
\title{Sample initial number of changepoints.}
\usage{
sampleK(mini, maxi, lambda, nb)
}
\arguments{
\item{mini}{Minimum value.}

\item{maxi}{Maximum value.}

\item{lambda}{Parameter of the Poisson distribution.}

\item{nb}{Number of values to sample.}
}
\value{
The sampled number of changepoints.
}
\description{
This function samples the initial number of changepoints from a sparse
Poisson prior.
}
\author{
Sophie Lebre
}
\references{
For more information on the prior choice and sampling, see:

Dondelinger et al. (2012), "Non-homogeneous dynamic Bayesian networks with
Bayesian regularization for inferring gene regulatory networks with
gradually time-varying structure", Machine Learning.
}

